+++ sources/API/Objects/Vector2.cpp
@@ -4,6 +4,6 @@ API::Vector2::Vector2(const LuaStack& args) {
     if (args.size() != 2)
         throw_error("[Vector2:new] Incorrect number of arguments!");
 
-    x = args.get<float>();
-    y = args.get<float>();
+    x = args.get<double>();
+    y = args.get<double>();
 }
